Fix types in security
[MAILPOET-2535]

+++ lib/Util/Security.php
@@ -21,7 +21,15 @@ class Security {
    */
   static function generateRandomString($length = 5) {
     $length = max(5, (int)$length);
-    $string = base_convert(bin2hex(random_bytes(ceil(3 * $length / 4))), 16, 36); // phpcs:ignore
+    $string = base_convert(
+      bin2hex(
+        random_bytes(
+          (int)ceil(3 * $length / 4)
+        )
+      ),
+      16,
+      36
+    ); // phpcs:ignore
     return substr($string, 0, $length);
   }
